feat(upload): add streaming upload for large video files with Bunny.net
- Introduce StreamingUploader client for efficient large file uploads using PUT
- Replace TUS uploader with streaming upload for files > 100MB
- Update video upload section to use StreamingUploader
- Remove TUS support and dependencies
- Update Bunny server to support streaming upload URLs
- Enhance progress tracking, speed calculation, and session management
- Document streaming upload method and Bunny.net compatibility

This new streaming upload method improves memory usage and progress tracking for large video uploads, aligning with Bunny.net Stream API requirements.
