+++ package.json
@@ -62,7 +62,6 @@
 		"recharts": "^3.0.2",
 		"sonner": "^2.0.3",
 		"tailwind-merge": "^3.1.0",
-		"tus-js-client": "^4.3.1",
 		"tw-animate-css": "^1.2.8",
 		"vaul": "^1.1.2",
 		"zod": "^3.24.2"
